refactor: Integrate openai compatible code based on the factory pattern (#217)
* Draft

* migrate to newstructure and mock test with pytest

* Test with unittest

* fix class

* trigger PR sync

* fix test

* fix mypy error

---------
