+++ tests/utils/test_openrouter_client_utils.py
@@ -35,7 +35,7 @@ class TestOpenRouterClient(unittest.TestCase):
             os.getenv("OPENROUTER_API_KEY"),
             1000,
             0.8,
-            7.0,
+            0.7,
             8,
             False,
             1,
@@ -51,7 +51,7 @@ class TestOpenRouterClient(unittest.TestCase):
             os.getenv("OPENROUTER_API_KEY"),
             1000,
             0.8,
-            7.0,
+            0.7,
             8,
             False,
             1,
@@ -72,7 +72,7 @@ class TestOpenRouterClient(unittest.TestCase):
             os.getenv("OPENROUTER_API_KEY"),
             1000,
             0.8,
-            7.0,
+            0.7,
             8,
             False,
             1,
@@ -93,7 +93,7 @@ class TestOpenRouterClient(unittest.TestCase):
             os.getenv("OPENROUTER_API_KEY"),
             1000,
             0.8,
-            7.0,
+            0.7,
             8,
             False,
             1,
